requirements_production.txt
Flask==2.3.3
python-dotenv==1.0.0
web3==6.11.3
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
Pillow==10.0.1
pandas==2.1.1
numpy==1.24.4
opencv-python-headless==4.8.1.78
cryptography==41.0.7
eth-account==0.9.0
eth-hash==0.5.2
hexbytes==0.3.1
jsonschema==4.19.1
lru-dict==1.2.0
parsimonious==0.9.0
protobuf==4.24.4
pyunormalize==15.0.0
rlp==3.0.0
typing-extensions==4.8.0
aiohttp==3.8.6
attrs==23.1.0
multidict==6.0.4
yarl==1.9.2
async-timeout==4.0.3
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.6
certifi==2023.7.22